+++ spring-cloud-netflix-core/src/main/java/org/springframework/cloud/netflix/zuul/endpoints/FiltersEndpoint.java
@@ -2,66 +2,55 @@ package org.springframework.cloud.netflix.zuul.endpoints;
 
 import com.netflix.zuul.ZuulFilter;
 import com.netflix.zuul.filters.FilterRegistry;
-import org.springframework.boot.actuate.endpoint.Endpoint;
-import org.springframework.boot.actuate.endpoint.mvc.MvcEndpoint;
+import org.springframework.boot.actuate.endpoint.AbstractEndpoint;
 import org.springframework.jmx.export.annotation.ManagedAttribute;
 import org.springframework.jmx.export.annotation.ManagedResource;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
 
 /**
- * Endpoint for listing Zuul filters
+ * Endpoint for listing Zuul filters.
+ *
+ * @author Daryl Robbins
+ * @author Gregor Zurowski
  */
 @ManagedResource(description = "List Zuul filters")
-public class FiltersEndpoint implements MvcEndpoint {
+public class FiltersEndpoint extends AbstractEndpoint<Map<String, List<Map<String, Object>>>> {
 
-    @RequestMapping(method = RequestMethod.GET)
-    @ResponseBody
+    private static final String ID = "filters";
+
+    public FiltersEndpoint() {
+        super(ID, true);
+    }
+
     @ManagedAttribute
-    public Map<String, List<Map<String, Object>>> getFilters() {
+    @Override
+    public Map<String, List<Map<String, Object>>> invoke() {
         final FilterRegistry filterRegistry = FilterRegistry.instance();
 
         // Map of filters by type
-        final Map<String, List<Map<String, Object>>> filterMap =
-                new TreeMap<String, List<Map<String, Object>>>();
+        final Map<String, List<Map<String, Object>>> filterMap = new TreeMap<>();
 
-        for (ZuulFilter f : filterRegistry.getAllFilters()) {
-
+        for (ZuulFilter filter : filterRegistry.getAllFilters()) {
             // Ensure that we have a list to store filters of each type
-            if (!filterMap.containsKey(f.filterType())) {
-                filterMap.put(f.filterType(), new ArrayList<Map<String, Object>>());
+            if (!filterMap.containsKey(filter.filterType())) {
+                filterMap.put(filter.filterType(), new ArrayList<Map<String, Object>>());
             }
 
+            final Map<String, Object> filterInfo = new LinkedHashMap<>();
+            filterInfo.put("class", filter.getClass().getName());
+            filterInfo.put("order", filter.filterOrder());
+            filterInfo.put("disabled", filter.isFilterDisabled());
+            filterInfo.put("static", filter.isStaticFilter());
 
-            final Map<String, Object> filterInfo = new LinkedHashMap<String, Object>();
-
-            filterInfo.put("class", f.getClass().getName());
-            filterInfo.put("order", f.filterOrder());
-            filterInfo.put("disabled", f.isFilterDisabled());
-            filterInfo.put("static", f.isStaticFilter());
-
-            filterMap.get(f.filterType()).add(filterInfo);
+            filterMap.get(filter.filterType()).add(filterInfo);
         }
 
         return filterMap;
-    }
-
-    @Override
-    public String getPath() {
-        return "/filters";
-    }
-
-    @Override
-    public boolean isSensitive() {
-        return true;
-    }
-
-    @Override
-    public Class<? extends Endpoint<?>> getEndpointType() {
-        return null;
     }
 
 }
